refactor(api-server): split routes/auth.rs into per-surface submodules
Behavior-preserving structural refactor of the 2950-line routes/auth.rs
churn hotspot. Extracts three cohesive, contiguous surfaces into focused
submodules under routes/auth/ without changing any route behavior,
handler signatures, or the public path surface:

  - registration.rs  -- register, verify_email, resend_verification (Story 1.1)
  - password_reset.rs -- forgot_password, reset_password (Story 1.4)
  - sessions.rs       -- list_sessions, revoke_session, revoke_all_sessions (Story 1.5)

Shared helpers (email/token/cookie rate-limit + auth utilities), the
login/refresh/logout cluster, the /me profile handlers, and the existing
test modules stay in auth/mod.rs. Submodules reach shared helpers via
`use super::*` (child modules can see parent-private items, so no helper
visibility changed). mod.rs re-exports each submodule with `pub use ...::*`
so `routes::auth::<name>` references (router, main.rs paths()/components())
keep resolving unchanged -- glob re-exports also carry utoipa `__path_*`
items. No SQL, no endpoint, no schema changes.
