+++ backend/servers/api-server/src/routes/auth/mod.rs
@@ -108,416 +108,17 @@ pub fn router() -> Router<AppState> {
         .route("/me", get(get_me).patch(update_me))
 }
 
-// ==================== Register (Story 1.1) ====================
+mod password_reset;
+mod registration;
+mod sessions;
 
-/// Register request.
-#[derive(Debug, Deserialize, ToSchema)]
-pub struct RegisterRequest {
-    /// Email address
-    pub email: String,
-    /// Password (min 8 characters, 1 uppercase, 1 number)
-    pub password: String,
-    /// Display name
-    pub name: String,
-    /// Phone number (optional)
-    pub phone: Option<String>,
-    /// Preferred locale (sk, cs, de, en)
-    pub locale: Option<String>,
-}
-
-/// Register response.
-///
-/// Intentionally generic and identical for both a fresh registration and an
-/// attempt to register an already-existing email (#956). It carries no
-/// account-specific data (notably no user id), so the response cannot be used
-/// to enumerate registered accounts.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct RegisterResponse {
-    /// Generic success message
-    pub message: String,
-}
-
-/// Generic, account-agnostic message returned by `register` regardless of
-/// whether the email was already registered. Keeps the success and
-/// already-exists paths byte-for-byte identical (#956).
-const REGISTER_GENERIC_MESSAGE: &str = "Check your email to verify your account";
-
-/// Register endpoint.
-#[utoipa::path(
-    post,
-    path = "/api/v1/auth/register",
-    tag = "Authentication",
-    request_body = RegisterRequest,
-    responses(
-        // A 201 with a generic body is returned whether or not the email is
-        // already registered, to avoid account enumeration (#956).
-        (status = 201, description = "Registration accepted", body = RegisterResponse),
-        (status = 400, description = "Invalid input", body = ErrorResponse)
-    )
-)]
-pub async fn register(
-    State(state): State<AppState>,
-    Json(req): Json<RegisterRequest>,
-) -> Result<impl IntoResponse, (StatusCode, Json<ErrorResponse>)> {
-    // Validate email format
-    if !AuthService::validate_email(&req.email) {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::new("INVALID_EMAIL", "Invalid email format")),
-        ));
-    }
-
-    // Validate password requirements
-    if let Err(errors) = AuthService::validate_password(&req.password) {
-        let details: Vec<ValidationError> = errors
-            .into_iter()
-            .map(|msg| ValidationError {
-                field: "password".to_string(),
-                message: msg.clone(),
-                code: "INVALID_PASSWORD".to_string(),
-            })
-            .collect();
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(
-                ErrorResponse::new("VALIDATION_ERROR", "Password does not meet requirements")
-                    .with_details(details),
-            ),
-        ));
-    }
-
-    // Validate name is not empty
-    if req.name.trim().is_empty() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::new("INVALID_NAME", "Name cannot be empty")),
-        ));
-    }
-
-    // Check if email already exists. We must NOT signal this back to the
-    // caller (no 409 / EMAIL_EXISTS) — a distinct response for an existing
-    // address is an account-enumeration oracle. Instead we short-circuit with
-    // the exact same generic 201 a fresh registration returns, and notify the
-    // real account holder out-of-band (logged here; an out-of-band "someone
-    // tried to register with your email" notification is a future enhancement).
-    match state.user_repo.email_exists(&req.email).await {
-        Ok(true) => {
-            tracing::info!(
-                email_hash = %common::email_log_hash(&req.email),
-                "Registration attempted for an already-registered email; returning generic response"
-            );
-            return Ok((
-                StatusCode::CREATED,
-                Json(RegisterResponse {
-                    message: REGISTER_GENERIC_MESSAGE.to_string(),
-                }),
-            ));
-        }
-        Ok(false) => {}
-        Err(e) => {
-            tracing::error!(error = %e, "Database error checking email");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to check email",
-                )),
-            ));
-        }
-    }
-
-    // Hash password
-    let password_hash = match state.auth_service.hash_password(&req.password) {
-        Ok(hash) => hash,
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to hash password");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "INTERNAL_ERROR",
-                    "Failed to process password",
-                )),
-            ));
-        }
-    };
-
-    // Determine locale from request or default to English
-    let locale = req
-        .locale
-        .as_ref()
-        .map(|l| Locale::parse(l))
-        .unwrap_or(Locale::English);
-
-    // Create user
-    let create_user = CreateUser {
-        email: req.email.clone(),
-        password_hash,
-        name: req.name.clone(),
-        phone: req.phone.clone(),
-        locale: locale.clone(),
-    };
-
-    let user = match state.user_repo.create(create_user).await {
-        Ok(user) => user,
-        Err(e) => {
-            tracing::error!(
-                error = %e,
-                email_hash = %common::email_log_hash(&req.email),
-                "Failed to create user"
-            );
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to create account",
-                )),
-            ));
-        }
-    };
-
-    // Generate verification token
-    let token = state.auth_service.generate_token();
-    let token_hash = state.auth_service.hash_token(&token);
-
-    // Store verification token
-    if let Err(e) = state
-        .user_repo
-        .create_verification_token(user.id, &token_hash)
-        .await
-    {
-        tracing::error!(error = %e, user_id = %user.id, "Failed to create verification token");
-        // Continue anyway - user can request resend
-    }
-
-    // Send verification email
-    if let Err(e) = state
-        .email_service
-        .send_verification_email(&user.email, &user.name, &token, &locale)
-        .await
-    {
-        tracing::error!(error = %e, user_id = %user.id, "Failed to send verification email");
-        // Continue anyway - user can request resend
-    }
-
-    tracing::info!(
-        user_id = %user.id,
-        "User registered successfully"
-    );
-
-    Ok((
-        StatusCode::CREATED,
-        Json(RegisterResponse {
-            message: REGISTER_GENERIC_MESSAGE.to_string(),
-        }),
-    ))
-}
-
-// ==================== Verify Email (Story 1.1) ====================
-
-/// Verify email query parameters.
-#[derive(Debug, Deserialize, ToSchema)]
-pub struct VerifyEmailQuery {
-    /// Verification token from email
-    pub token: String,
-}
-
-/// Verify email response.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct VerifyEmailResponse {
-    /// Success message
-    pub message: String,
-}
-
-/// Verify email endpoint.
-#[utoipa::path(
-    get,
-    path = "/api/v1/auth/verify-email",
-    tag = "Authentication",
-    params(
-        ("token" = String, Query, description = "Email verification token")
-    ),
-    responses(
-        (status = 200, description = "Email verified", body = VerifyEmailResponse),
-        (status = 400, description = "Invalid or expired token", body = ErrorResponse)
-    )
-)]
-pub async fn verify_email(
-    State(state): State<AppState>,
-    axum::extract::Query(query): axum::extract::Query<VerifyEmailQuery>,
-) -> Result<Json<VerifyEmailResponse>, (StatusCode, Json<ErrorResponse>)> {
-    // Hash the token to look it up
-    let token_hash = state.auth_service.hash_token(&query.token);
-
-    // Find the token
-    let verification_token = match state.user_repo.find_verification_token(&token_hash).await {
-        Ok(Some(token)) => token,
-        Ok(None) => {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::new(
-                    "INVALID_TOKEN",
-                    "This verification link is invalid or has already been used",
-                )),
-            ));
-        }
-        Err(e) => {
-            tracing::error!(error = %e, "Database error finding verification token");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to verify token",
-                )),
-            ));
-        }
-    };
-
-    // Check if token is expired
-    if verification_token.is_expired() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::new(
-                "TOKEN_EXPIRED",
-                "This verification link has expired. Please request a new one.",
-            )),
-        ));
-    }
-
-    // Mark token as used
-    if let Err(e) = state
-        .user_repo
-        .use_verification_token(verification_token.id)
-        .await
-    {
-        tracing::error!(error = %e, "Failed to mark verification token as used");
-    }
-
-    // Verify the user's email
-    match state
-        .user_repo
-        .verify_email(verification_token.user_id)
-        .await
-    {
-        Ok(Some(user)) => {
-            tracing::info!(user_id = %user.id, "Email verified");
-            Ok(Json(VerifyEmailResponse {
-                message: "Your email has been verified. You can now log in.".to_string(),
-            }))
-        }
-        Ok(None) => Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::new(
-                "USER_NOT_FOUND",
-                "User account not found or already verified",
-            )),
-        )),
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to verify email");
-            Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to verify email",
-                )),
-            ))
-        }
-    }
-}
-
-// ==================== Resend Verification (Story 1.1) ====================
-
-/// Resend verification request.
-#[derive(Debug, Deserialize, ToSchema)]
-pub struct ResendVerificationRequest {
-    /// Email address
-    pub email: String,
-}
-
-/// Resend verification response.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ResendVerificationResponse {
-    /// Success message
-    pub message: String,
-}
-
-/// Resend verification email endpoint.
-#[utoipa::path(
-    post,
-    path = "/api/v1/auth/resend-verification",
-    tag = "Authentication",
-    request_body = ResendVerificationRequest,
-    responses(
-        (status = 200, description = "Verification email sent (if account exists)", body = ResendVerificationResponse),
-        (status = 429, description = "Too many requests for this email", body = ErrorResponse),
-    )
-)]
-pub async fn resend_verification(
-    State(state): State<AppState>,
-    Json(req): Json<ResendVerificationRequest>,
-) -> Result<Json<ResendVerificationResponse>, (StatusCode, Json<ErrorResponse>)> {
-    // Throttle per email before doing any work: this endpoint sends an email
-    // and rotates the recipient's verification token on every call, so it is a
-    // mailbomb / token-clobber surface (see limiter docs above).
-    if !resend_verification_rate_allowed(&req.email) {
-        tracing::warn!(
-            email_hash = %common::email_log_hash(&req.email),
-            "resend-verification blocked due to rate limiting"
-        );
-        return Err(email_dispatch_rate_limited());
-    }
-
-    // Always return success to prevent email enumeration
-    let response = ResendVerificationResponse {
-        message:
-            "If an unverified account exists with this email, a verification link has been sent."
-                .to_string(),
-    };
-
-    // Try to find user and send email
-    match state.user_repo.find_by_email(&req.email).await {
-        Ok(Some(user)) if user.status == "pending" => {
-            // Invalidate old tokens
-            let _ = state.user_repo.invalidate_user_tokens(user.id).await;
-
-            // Generate new token
-            let token = state.auth_service.generate_token();
-            let token_hash = state.auth_service.hash_token(&token);
-
-            // Store new token
-            if let Err(e) = state
-                .user_repo
-                .create_verification_token(user.id, &token_hash)
-                .await
-            {
-                tracing::error!(error = %e, user_id = %user.id, "Failed to create verification token");
-                return Ok(Json(response));
-            }
-
-            // Send email
-            if let Err(e) = state
-                .email_service
-                .send_verification_email(&user.email, &user.name, &token, &user.locale_enum())
-                .await
-            {
-                tracing::error!(error = %e, user_id = %user.id, "Failed to send verification email");
-            }
-
-            tracing::info!(user_id = %user.id, "Resent verification email");
-        }
-        _ => {
-            // User not found or already verified - don't reveal this
-            tracing::debug!(
-                email_hash = %common::email_log_hash(&req.email),
-                "Resend verification request for non-pending account"
-            );
-        }
-    }
-
-    Ok(Json(response))
-}
+// Re-export each submodule's public handlers + DTO types so `routes::auth::<name>`
+// references keep resolving unchanged after the per-surface split: the `router()`
+// above, plus `paths(...)` / `components(schemas(...))` in `main.rs`. Glob
+// re-exports also carry the utoipa-generated `__path_*` items next to each fn.
+pub use password_reset::*;
+pub use registration::*;
+pub use sessions::*;
 
 // ==================== Login (Story 1.2) ====================
 
@@ -1591,600 +1192,6 @@ pub async fn logout(
         }),
     )
         .into_response())
-}
-
-// ==================== Forgot Password (Story 1.4) ====================
-
-/// Forgot password request.
-#[derive(Debug, Deserialize, ToSchema)]
-pub struct ForgotPasswordRequest {
-    /// Email address
-    pub email: String,
-}
-
-/// Forgot password response.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ForgotPasswordResponse {
-    /// Success message (always same to prevent enumeration)
-    pub message: String,
-}
-
-/// Forgot password endpoint - initiates password reset.
-#[utoipa::path(
-    post,
-    path = "/api/v1/auth/forgot-password",
-    tag = "Authentication",
-    request_body = ForgotPasswordRequest,
-    responses(
-        (status = 200, description = "Password reset email sent (if account exists)", body = ForgotPasswordResponse),
-        (status = 429, description = "Too many requests for this email", body = ErrorResponse),
-    )
-)]
-pub async fn forgot_password(
-    State(state): State<AppState>,
-    Json(req): Json<ForgotPasswordRequest>,
-) -> Result<Json<ForgotPasswordResponse>, (StatusCode, Json<ErrorResponse>)> {
-    // Throttle per email before doing any work: this endpoint sends an email
-    // and invalidates the recipient's outstanding reset token on every call, so
-    // it is a mailbomb / token-clobber surface (see limiter docs above).
-    if !forgot_password_rate_allowed(&req.email) {
-        tracing::warn!(
-            email_hash = %common::email_log_hash(&req.email),
-            "forgot-password blocked due to rate limiting"
-        );
-        return Err(email_dispatch_rate_limited());
-    }
-
-    // Always return success to prevent email enumeration
-    let response = ForgotPasswordResponse {
-        message: "If an account exists with this email, a password reset link has been sent."
-            .to_string(),
-    };
-
-    // Try to find user and send email
-    match state.user_repo.find_by_email(&req.email).await {
-        Ok(Some(user)) if user.status == "active" => {
-            // Invalidate any existing reset tokens for this user
-            let _ = state
-                .password_reset_repo
-                .invalidate_user_tokens(user.id)
-                .await;
-
-            // Generate reset token (1 hour expiry)
-            let token = state.auth_service.generate_token();
-            let token_hash = state.auth_service.hash_token(&token);
-            let expires_at = chrono::Utc::now() + chrono::Duration::hours(1);
-
-            // Store token
-            use db::models::CreatePasswordResetToken;
-            let create_token = CreatePasswordResetToken {
-                user_id: user.id,
-                token_hash,
-                expires_at,
-            };
-
-            if let Err(e) = state.password_reset_repo.create(create_token).await {
-                tracing::error!(error = %e, user_id = %user.id, "Failed to create password reset token");
-                return Ok(Json(response));
-            }
-
-            // Send reset email
-            if let Err(e) = state
-                .email_service
-                .send_password_reset_email(&user.email, &user.name, &token, &user.locale_enum())
-                .await
-            {
-                tracing::error!(error = %e, user_id = %user.id, "Failed to send password reset email");
-            }
-
-            tracing::info!(user_id = %user.id, "Password reset email sent");
-        }
-        Ok(Some(user)) => {
-            // User exists but not active (pending/suspended/deleted)
-            tracing::debug!(
-                user_id = %user.id,
-                status = %user.status,
-                "Password reset requested for non-active account"
-            );
-        }
-        Ok(None) => {
-            // User not found - don't reveal this
-            tracing::debug!(
-                email_hash = %common::email_log_hash(&req.email),
-                "Password reset requested for unknown email"
-            );
-        }
-        Err(e) => {
-            tracing::error!(error = %e, "Database error finding user for password reset");
-        }
-    }
-
-    Ok(Json(response))
-}
-
-// ==================== Reset Password (Story 1.4) ====================
-
-/// Reset password request.
-#[derive(Debug, Deserialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ResetPasswordRequest {
-    /// Reset token from email
-    pub token: String,
-    /// New password (min 8 characters, 1 uppercase, 1 number)
-    pub new_password: String,
-}
-
-/// Reset password response.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ResetPasswordResponse {
-    /// Success message
-    pub message: String,
-}
-
-/// Reset password endpoint - completes password reset.
-#[utoipa::path(
-    post,
-    path = "/api/v1/auth/reset-password",
-    tag = "Authentication",
-    request_body = ResetPasswordRequest,
-    responses(
-        (status = 200, description = "Password reset successful", body = ResetPasswordResponse),
-        (status = 400, description = "Invalid or expired token", body = ErrorResponse),
-        (status = 400, description = "Password does not meet requirements", body = ErrorResponse)
-    )
-)]
-pub async fn reset_password(
-    State(state): State<AppState>,
-    Json(req): Json<ResetPasswordRequest>,
-) -> Result<Json<ResetPasswordResponse>, (StatusCode, Json<ErrorResponse>)> {
-    // Validate new password requirements
-    if let Err(errors) = AuthService::validate_password(&req.new_password) {
-        let details: Vec<ValidationError> = errors
-            .into_iter()
-            .map(|msg| ValidationError {
-                field: "new_password".to_string(),
-                message: msg.clone(),
-                code: "INVALID_PASSWORD".to_string(),
-            })
-            .collect();
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(
-                ErrorResponse::new("VALIDATION_ERROR", "Password does not meet requirements")
-                    .with_details(details),
-            ),
-        ));
-    }
-
-    // Hash the token to look it up
-    let token_hash = state.auth_service.hash_token(&req.token);
-
-    // Find the reset token
-    let reset_token = match state
-        .password_reset_repo
-        .find_by_token_hash(&token_hash)
-        .await
-    {
-        Ok(Some(token)) => token,
-        Ok(None) => {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::new(
-                    "INVALID_TOKEN",
-                    "This password reset link is invalid or has already been used",
-                )),
-            ));
-        }
-        Err(e) => {
-            tracing::error!(error = %e, "Database error finding password reset token");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to validate reset token",
-                )),
-            ));
-        }
-    };
-
-    // Check if token is expired
-    if reset_token.is_expired() {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::new(
-                "TOKEN_EXPIRED",
-                "This password reset link has expired. Please request a new one.",
-            )),
-        ));
-    }
-
-    // Find the user
-    let user = match state.user_repo.find_by_id(reset_token.user_id).await {
-        Ok(Some(user)) => user,
-        Ok(None) => {
-            return Err((
-                StatusCode::BAD_REQUEST,
-                Json(ErrorResponse::new(
-                    "USER_NOT_FOUND",
-                    "User account no longer exists",
-                )),
-            ));
-        }
-        Err(e) => {
-            tracing::error!(error = %e, "Database error finding user");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to reset password",
-                )),
-            ));
-        }
-    };
-
-    // Check if user is still active
-    if user.status != "active" {
-        return Err((
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::new(
-                "ACCOUNT_INACTIVE",
-                "Cannot reset password for inactive account",
-            )),
-        ));
-    }
-
-    // D2.2: validate the new password against the user's effective per-org
-    // auth policy (strictest across active memberships; falls back to the
-    // platform default for users with no orgs). The `AuthService::validate_password`
-    // call above enforces the platform-default rules; this enforcer call adds
-    // any per-org tightening (longer min length, required char classes).
-    let enforcer = crate::services::AuthPolicyEnforcer::new(state.db.clone());
-    if let Err(err) = enforcer
-        .check_password_change(user.id, &req.new_password)
-        .await
-    {
-        match err {
-            crate::services::AuthPolicyError::PasswordPolicy(violations) => {
-                let details: Vec<ValidationError> = violations
-                    .into_iter()
-                    .map(|msg| ValidationError {
-                        field: "new_password".to_string(),
-                        message: msg,
-                        code: "INVALID_PASSWORD".to_string(),
-                    })
-                    .collect();
-                return Err((
-                    StatusCode::UNPROCESSABLE_ENTITY,
-                    Json(
-                        ErrorResponse::new(
-                            "PASSWORD_POLICY_VIOLATION",
-                            "Password does not satisfy org policy",
-                        )
-                        .with_details(details),
-                    ),
-                ));
-            }
-            other => {
-                tracing::error!(error = %other, user_id = %user.id, "Password policy check failed");
-                return Err((
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                    Json(ErrorResponse::new(
-                        "AUTH_POLICY_ERROR",
-                        "Failed to validate password",
-                    )),
-                ));
-            }
-        }
-    }
-
-    // Hash new password
-    let password_hash = match state.auth_service.hash_password(&req.new_password) {
-        Ok(hash) => hash,
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to hash new password");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "INTERNAL_ERROR",
-                    "Failed to process password",
-                )),
-            ));
-        }
-    };
-
-    // Update password
-    if let Err(e) = state
-        .user_repo
-        .update_password(user.id, &password_hash)
-        .await
-    {
-        tracing::error!(error = %e, user_id = %user.id, "Failed to update password");
-        return Err((
-            StatusCode::INTERNAL_SERVER_ERROR,
-            Json(ErrorResponse::new(
-                "DATABASE_ERROR",
-                "Failed to update password",
-            )),
-        ));
-    }
-
-    // Mark token as used
-    if let Err(e) = state.password_reset_repo.mark_used(reset_token.id).await {
-        tracing::error!(error = %e, "Failed to mark reset token as used");
-        // Continue anyway - password was changed successfully
-    }
-
-    // Revoke all refresh tokens for security (force re-login)
-    if let Err(e) = state
-        .session_repo
-        .revoke_all_user_tokens(user.id, None)
-        .await
-    {
-        tracing::error!(error = %e, "Failed to revoke user sessions");
-        // Continue anyway - password was changed
-    }
-
-    tracing::info!(user_id = %user.id, "Password reset successfully");
-
-    Ok(Json(ResetPasswordResponse {
-        message: "Password has been reset. Please log in with your new password.".to_string(),
-    }))
-}
-
-// ==================== Session Management (Story 1.5) ====================
-
-/// Session info returned to clients.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct SessionInfo {
-    /// Session ID
-    pub id: String,
-    /// Device info (if available)
-    pub device_info: Option<String>,
-    /// IP address
-    pub ip_address: Option<String>,
-    /// User agent
-    pub user_agent: Option<String>,
-    /// When the session was created
-    pub created_at: String,
-    /// When the session was last used
-    pub last_used_at: String,
-    /// Whether this is the current session
-    pub is_current: bool,
-}
-
-/// List sessions response.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct ListSessionsResponse {
-    /// Active sessions
-    pub sessions: Vec<SessionInfo>,
-}
-
-/// List active sessions endpoint.
-#[utoipa::path(
-    get,
-    path = "/api/v1/auth/sessions",
-    tag = "Authentication",
-    security(("bearer_auth" = [])),
-    responses(
-        (status = 200, description = "Sessions retrieved", body = ListSessionsResponse),
-        (status = 401, description = "Not authenticated", body = ErrorResponse)
-    )
-)]
-pub async fn list_sessions(
-    State(state): State<AppState>,
-    headers: axum::http::HeaderMap,
-) -> Result<Json<ListSessionsResponse>, (StatusCode, Json<ErrorResponse>)> {
-    let user_id = authenticated_user_id(&state, &headers)?;
-
-    // Identify the caller's current session. Prefer the HttpOnly cookie,
-    // fall back to the `X-Refresh-Token` header (see
-    // `resolve_current_session_id`). Cookie-based ppt-web clients used to
-    // resolve to `None` here, so no row could ever be marked `isCurrent`.
-    let current_session_id = resolve_current_session_id(&state, &headers, user_id).await;
-
-    // Get all active sessions for user
-    let sessions = match state.session_repo.find_user_sessions(user_id).await {
-        Ok(sessions) => sessions,
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to fetch user sessions");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to fetch sessions",
-                )),
-            ));
-        }
-    };
-
-    let session_infos: Vec<SessionInfo> = sessions
-        .into_iter()
-        .map(|s| {
-            let is_current = current_session_id
-                .as_ref()
-                .map(|id| id == &s.id)
-                .unwrap_or(false);
-
-            SessionInfo {
-                id: s.id.to_string(),
-                device_info: s.device_info,
-                ip_address: s.ip_address,
-                user_agent: s.user_agent,
-                created_at: s.created_at.to_rfc3339(),
-                last_used_at: s.last_used_at.to_rfc3339(),
-                is_current,
-            }
-        })
-        .collect();
-
-    Ok(Json(ListSessionsResponse {
-        sessions: session_infos,
-    }))
-}
-
-/// Revoke session request.
-#[derive(Debug, Deserialize, ToSchema)]
-pub struct RevokeSessionRequest {
-    /// Session ID to revoke
-    pub session_id: String,
-}
-
-/// Revoke session response.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct RevokeSessionResponse {
-    /// Success message
-    pub message: String,
-}
-
-/// Revoke a specific session.
-#[utoipa::path(
-    post,
-    path = "/api/v1/auth/sessions/revoke",
-    tag = "Authentication",
-    security(("bearer_auth" = [])),
-    request_body = RevokeSessionRequest,
-    responses(
-        (status = 200, description = "Session revoked", body = RevokeSessionResponse),
-        (status = 401, description = "Not authenticated", body = ErrorResponse),
-        (status = 404, description = "Session not found", body = ErrorResponse)
-    )
-)]
-pub async fn revoke_session(
-    State(state): State<AppState>,
-    headers: axum::http::HeaderMap,
-    Json(req): Json<RevokeSessionRequest>,
-) -> Result<Json<RevokeSessionResponse>, (StatusCode, Json<ErrorResponse>)> {
-    let user_id = authenticated_user_id(&state, &headers)?;
-
-    // Parse session ID
-    let session_id: uuid::Uuid = req.session_id.parse().map_err(|_| {
-        (
-            StatusCode::BAD_REQUEST,
-            Json(ErrorResponse::new(
-                "INVALID_SESSION_ID",
-                "Invalid session ID format",
-            )),
-        )
-    })?;
-
-    // Verify session belongs to this user
-    let sessions = match state.session_repo.find_user_sessions(user_id).await {
-        Ok(sessions) => sessions,
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to fetch user sessions");
-            return Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to verify session",
-                )),
-            ));
-        }
-    };
-
-    let session_exists = sessions.iter().any(|s| s.id == session_id);
-    if !session_exists {
-        return Err((
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse::new("SESSION_NOT_FOUND", "Session not found")),
-        ));
-    }
-
-    // Revoke the session
-    match state.session_repo.revoke_token(session_id).await {
-        Ok(true) => {
-            tracing::info!(user_id = %user_id, session_id = %session_id, "Session revoked");
-            Ok(Json(RevokeSessionResponse {
-                message: "Session revoked successfully".to_string(),
-            }))
-        }
-        Ok(false) => Err((
-            StatusCode::NOT_FOUND,
-            Json(ErrorResponse::new(
-                "SESSION_NOT_FOUND",
-                "Session already revoked",
-            )),
-        )),
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to revoke session");
-            Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to revoke session",
-                )),
-            ))
-        }
-    }
-}
-
-/// Revoke all sessions response.
-#[derive(Debug, Serialize, ToSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct RevokeAllSessionsResponse {
-    /// Success message
-    pub message: String,
-    /// Number of sessions revoked
-    pub revoked_count: u64,
-}
-
-/// Revoke all sessions except current.
-#[utoipa::path(
-    post,
-    path = "/api/v1/auth/sessions/revoke-all",
-    tag = "Authentication",
-    security(("bearer_auth" = [])),
-    responses(
-        (status = 200, description = "Sessions revoked", body = RevokeAllSessionsResponse),
-        (status = 401, description = "Not authenticated", body = ErrorResponse)
-    )
-)]
-pub async fn revoke_all_sessions(
-    State(state): State<AppState>,
-    headers: axum::http::HeaderMap,
-) -> Result<Json<RevokeAllSessionsResponse>, (StatusCode, Json<ErrorResponse>)> {
-    let user_id = authenticated_user_id(&state, &headers)?;
-
-    // Get current session to exclude. Prefer the HttpOnly cookie, fall back
-    // to the `X-Refresh-Token` header (see `resolve_current_session_id`).
-    // Cookie-based ppt-web clients used to resolve to `None` here, which made
-    // `revoke_all_user_tokens(user_id, None)` revoke the caller's OWN live
-    // session — "sign out other devices" signed the caller out too.
-    let current_session_id = resolve_current_session_id(&state, &headers, user_id).await;
-
-    // Revoke all sessions except current
-    match state
-        .session_repo
-        .revoke_all_user_tokens(user_id, current_session_id)
-        .await
-    {
-        Ok(count) => {
-            tracing::info!(
-                user_id = %user_id,
-                revoked_count = count,
-                "All other sessions revoked"
-            );
-            Ok(Json(RevokeAllSessionsResponse {
-                message: format!("{} session(s) revoked", count),
-                revoked_count: count,
-            }))
-        }
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to revoke sessions");
-            Err((
-                StatusCode::INTERNAL_SERVER_ERROR,
-                Json(ErrorResponse::new(
-                    "DATABASE_ERROR",
-                    "Failed to revoke sessions",
-                )),
-            ))
-        }
-    }
 }
 
 // ==================== Current User Profile (TypeSpec `/me`) ====================
